Refactor arm so its always PIDing.
