--- src/main/java/frc/robot/commands/Arm/HoldCommand.java
@@ -1,34 +1,0 @@
-package frc.robot.commands.Arm;
-
-import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.Constants.Arm;
-import frc.robot.subsystems.Arm.ArmSubsystem;
-
-public class HoldCommand extends Command {
-    private ArmSubsystem armSubsystem;
-
-    public HoldCommand(ArmSubsystem armSubsystem) {
-        this.armSubsystem = armSubsystem;
-        addRequirements(armSubsystem);
-    }
-
-    @Override
-    public void initialize() {
-        armSubsystem.setAngleToCurrent();
-    }
-
-    @Override
-    public void execute() {
-        armSubsystem.execute();
-    }
-
-    @Override
-    public boolean isFinished() {
-        return false;
-    }
-
-    @Override
-    public void end(boolean interrupted) {
-        // armSubsystem.setVoltage(0);
-    }
-}
